+++ lib/private/legacy/OC_Helper.php
@@ -106,65 +106,6 @@ class OC_Helper {
 	}
 
 	/**
-	 * Adds a suffix to the name in case the file exists
-	 *
-	 * @param string $path
-	 * @param string $filename
-	 * @return string
-	 */
-	public static function buildNotExistingFileName($path, $filename) {
-		$view = \OC\Files\Filesystem::getView();
-		return self::buildNotExistingFileNameForView($path, $filename, $view);
-	}
-
-	/**
-	 * Adds a suffix to the name in case the file exists
-	 *
-	 * @param string $path
-	 * @param string $filename
-	 * @return string
-	 */
-	public static function buildNotExistingFileNameForView($path, $filename, \OC\Files\View $view) {
-		if ($path === '/') {
-			$path = '';
-		}
-		if ($pos = strrpos($filename, '.')) {
-			$name = substr($filename, 0, $pos);
-			$ext = substr($filename, $pos);
-		} else {
-			$name = $filename;
-			$ext = '';
-		}
-
-		$newpath = $path . '/' . $filename;
-		if ($view->file_exists($newpath)) {
-			if (preg_match_all('/\((\d+)\)/', $name, $matches, PREG_OFFSET_CAPTURE)) {
-				//Replace the last "(number)" with "(number+1)"
-				$last_match = count($matches[0]) - 1;
-				$counter = $matches[1][$last_match][0] + 1;
-				$offset = $matches[0][$last_match][1];
-				$match_length = strlen($matches[0][$last_match][0]);
-			} else {
-				$counter = 2;
-				$match_length = 0;
-				$offset = false;
-			}
-			do {
-				if ($offset) {
-					//Replace the last "(number)" with "(number+1)"
-					$newname = substr_replace($name, '(' . $counter . ')', $offset, $match_length);
-				} else {
-					$newname = $name . ' (' . $counter . ')';
-				}
-				$newpath = $path . '/' . $newname . $ext;
-				$counter++;
-			} while ($view->file_exists($newpath));
-		}
-
-		return $newpath;
-	}
-
-	/**
 	 * Checks if a function is available
 	 *
 	 * @deprecated 25.0.0 use \OCP\Util::isFunctionEnabled instead
