+++ lib/public/Files.php
@@ -126,16 +126,4 @@ class Files {
 		}
 		return $includeResult ? [$count, $result] : $count;
 	}
-
-	/**
-	 * Adds a suffix to the name in case the file exists
-	 * @param string $path
-	 * @param string $filename
-	 * @return string
-	 * @since 5.0.0
-	 * @deprecated 14.0.0 use getNonExistingName of the OCP\Files\Folder object
-	 */
-	public static function buildNotExistingFileName($path, $filename) {
-		return \OC_Helper::buildNotExistingFileName($path, $filename);
-	}
 }
